Renames imageElement to DemoImageElement
