+++ demo/index.ts
@@ -6,7 +6,7 @@ import { Schema } from "prosemirror-model";
 import { schema as basicSchema } from "prosemirror-schema-basic";
 import { EditorState } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
-import { createImageElement } from "../src/elements/image/imageElement";
+import { createImageElement } from "../src/elements/demo-image/DemoImageElement";
 import { buildElementPlugin } from "../src/plugin/element";
 import {
   createParsers,
